refactor(listener): extract cron commands
